+++ src/app/AppSettings.h
@@ -15,6 +15,7 @@ class AppSettings final : public QObject {
   Q_PROPERTY(bool steamEnabled READ steamEnabled WRITE setSteamEnabled NOTIFY sourcesChanged)
   Q_PROPERTY(bool lutrisEnabled READ lutrisEnabled WRITE setLutrisEnabled NOTIFY sourcesChanged)
   Q_PROPERTY(bool heroicEnabled READ heroicEnabled WRITE setHeroicEnabled NOTIFY sourcesChanged)
+  Q_PROPERTY(bool faugusEnabled READ faugusEnabled WRITE setFaugusEnabled NOTIFY sourcesChanged)
   Q_PROPERTY(bool closeAfterLaunch READ closeAfterLaunch WRITE setCloseAfterLaunch NOTIFY
                  closeAfterLaunchChanged)
 
@@ -35,6 +36,8 @@ public:
   void setLutrisEnabled(bool value);
   [[nodiscard]] bool heroicEnabled() const;
   void setHeroicEnabled(bool value);
+  [[nodiscard]] bool faugusEnabled() const;
+  void setFaugusEnabled(bool value);
   [[nodiscard]] bool closeAfterLaunch() const;
   void setCloseAfterLaunch(bool value);
 
@@ -59,5 +62,6 @@ private:
   bool m_steamEnabled = true;
   bool m_lutrisEnabled = true;
   bool m_heroicEnabled = true;
+  bool m_faugusEnabled = true;
   bool m_closeAfterLaunch = false;
 };